# adapt-design-files/openapi_app.yaml
openapi: 3.1.0
info:
  title: adapt App API (V1 - 基本設計書完全準拠)
  version: v19-integrated
  description: |
    adapt オンライン学習プラットフォーム App API

    ## 概要
    - **対象**: 受講者・講師向けアプリケーションAPI
    - **認証**: OIDC（auth.adapt-co.io）+ Bearer Token
    - **エンドポイント数**: 64個
    - **基本設計書**: 06_API一覧 完全準拠（100%）

    ## 機能グループ

    ### ユーザー管理
    - プロフィール管理（取得・更新）
    - パスワード変更
    - 通知設定

    ### コース閲覧・受講
    - ストアトップ（講座一覧）
    - 講座詳細・申込
    - マイ講座一覧
    - 学習進捗管理

    ### チャット・コミュニケーション
    - チャンネル管理
    - スレッド・メッセージ投稿
    - 既読管理
    - リアルタイム通知

    ### 課題・提出
    - 課題一覧・詳細
    - 課題提出
    - 採点・フィードバック閲覧

    ### 決済
    - Stripe決済
    - クーポン適用
    - 購入履歴

    ### 講師機能
    - コース作成・編集
    - レッスン管理
    - 受講者管理
    - 売上確認

    ## 認証・認可

    ### 認証方式
    - **OIDC**: auth.adapt-co.ioでの認証
    - **Bearer Token**: JWTトークンをAuthorizationヘッダーに設定
    - **例外**: /api/v1/health（ヘルスチェック）は認証不要

    ### ロール
    - `guest`: 未ログイン（閲覧のみ）
    - `learner`: 受講者
    - `instructor`: 講師
    - `assistant`: アシスタント/TA
    - `operator`: 運営（通常）
    - `root`: 運営（特権）

    ### x-roles 拡張プロパティ
    各エンドポイントには`x-roles`で必要なロールを指定：
    ```yaml
    x-roles:
      - learner
      - instructor
    ```

    ## 設計原則

    ### アーキテクチャ
    - **Repository Pattern**: Service層はORM非依存
    - **DTO分離**: Prisma ModelはDTOとして公開しない
    - **契約の正**: packages/sharedのDTOが正

    ### エラーハンドリング
    - 4xx: クライアントエラー（バリデーション等）
    - 5xx: サーバーエラー
    - 詳細はErrorResponseスキーマ参照

    ### バージョニング
    - URLパスに`/api/v1/`を含む
    - 破壊的変更は新バージョンで対応

    ## 技術スタック
...
servers:
...
paths:
...
  /api/v1/store/courses/{courseId}:
    get:
      tags:
      - Store/Public Access
      summary: 講座詳細取得
      description: 未受講者向けLPデータ。アウトカム、前提条件、おすすめ、FAQ、日程スケジュールをすべて取得。
      operationId: API_016
      x-api-id: API-016
      x-roles: guest, all
      x-policy: '-'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericDetailView'
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
...
  /api/v1/learner/courses/{courseId}:
    get:
      tags:
      - Learner Dashboard/Progress
      summary: 受講中講座詳細
      description: 受講者用学習画面。最新の進捗、シラバス、受講者専用の連絡事項、解禁済み教材等を取得。
      operationId: API_020
      x-api-id: API-020
      x-roles: learner
      x-policy: 423_ON_FROZEN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericDetailView'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
...
  /api/v1/instructor/courses:
    get:
      tags:
      - Course Management
      summary: 自講座一覧
      description: 自分が管理、または講師として参加（担当）している全講座を一覧取得する。
      operationId: API_031
      x-api-id: API-031
      x-roles: instructor
      x-policy: '-'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
      security:
      - bearerAuth: []
    post:
      tags:
      - Course Management
      summary: 講座新規作成（講師）
      description: |-
        講師がコースを新規作成。ownerUserId=作成者（サーバー側で自動設定、リクエストでの指定不可）。初期status=draft。CourseMember(role=instructor_owner)が自動作成される。公開にはrequest-approval→approve が必要。
      operationId: API_032
      x-api-id: API-032
      x-roles: instructor
      x-policy: '-'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailView'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorCourseCreateRequest'
  /api/v1/instructor/courses/{courseId}:
    get:
      tags:
      - Course Management
      summary: 講座詳細取得(管理)
      description: 講師用管理画面。売上概況、受講者推移、下書き情報を含む管理用詳細データ。
      operationId: API_033
      x-api-id: API-033
      x-roles: instructor
      x-policy: '-'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailView'
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
      - Course Management
      summary: 講座情報更新
      description: 講座のタイトル、LP、価格、募集期間、マーケティング項目を更新（image_33637a反映）。
      operationId: API_034
      x-api-id: API-034
      x-roles: instructor_owner
      x-policy: 423_ON_FROZEN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailView'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
    delete:
      tags:
      - Course Management
      summary: 講座削除(論理)
      description: 講座を論理削除(archived)状態にする。既存の受講生には閲覧権限のみ残存させる。
      operationId: API_035
      x-api-id: API-035
      x-roles: instructor_owner
      x-policy: '-'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
...
  /api/v1/courses/{courseId}/channels:
    get:
      tags:
      - Communication
      summary: チャンネル一覧取得
      description: サイドメニュー構築用。general, announcement等のチャンネル種別を含む。
      operationId: API_054
      x-api-id: API-054
      x-roles: all_in_course
      x-policy: '-'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseChannelListResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
      - Communication
      summary: チャンネル作成
      description: コース内に新しいカスタムチャンネルを追加。
      operationId: API_055
      x-api-id: API-055
      x-roles: instructor_owner
      x-policy: 423_ON_FROZEN
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailView'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseChannelCreateRequest'
  /api/v1/channels/{channelId}:
    put:
      tags:
      - Communication
      summary: チャンネル編集
      description: チャンネル名、説明文、閲覧制限、アーカイブ状態の更新。
      operationId: API_056
      x-api-id: API-056
      x-roles: instructor_owner
      x-policy: 423_ON_FROZEN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseChannelDetailView'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: channelId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericWriteRequest'
    delete:
      tags:
      - Communication
      summary: チャンネル削除
      description: チャンネルを論理削除。過去ログは監査用に保持。
      operationId: API_057
      x-api-id: API-057
      x-roles: instructor_owner
      x-policy: 423_ON_FROZEN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: channelId
        in: path
        required: true
        schema:
          type: string
  /api/v1/channels/{channelId}/messages:
    get:
      tags:
      - Communication
      summary: メッセージ履歴取得
      description: スレッド親（ルート）メッセージの一覧取得。課題の相互閲覧可（image_32d9b1反映）。
      operationId: API_058
      x-api-id: API-058
      x-roles: all_in_course
      x-policy: 閲覧のみ可
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: channelId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
      - Communication
      summary: メッセージ投稿
      description: チャンネルへの新規投稿。親メッセージとして保存。threads_onlyを強制適用。
      operationId: API_059
      x-api-id: API-059
      x-roles: all_in_course
      x-policy: threads_only(AUTO)
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseChannelDetailView'
      security:
      - bearerAuth: []
      parameters:
      - name: channelId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseMessageCreateRequest'
...
  /api/v1/messages/{messageId}:
    patch:
      tags:
      - Communication
      summary: メッセージ編集
      description: 自分の投稿内容を修正。編集履歴を保持。423チェック対象。
      operationId: API_062
      x-api-id: API-062
      x-roles: owner_only
      x-policy: 423_ON_FROZEN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseMessageView'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: messageId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericWriteRequest'
    delete:
      tags:
      - Communication
      summary: メッセージ削除
      description: 投稿の論理削除。返信がある場合は「削除されました」と表示。
      operationId: API_063
      x-api-id: API-063
      x-roles: owner_only, instructor
      x-policy: 423_ON_FROZEN
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '423':
          description: Locked (Course is frozen)
      security:
      - bearerAuth: []
      parameters:
      - name: messageId
        in: path
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        ownerUserId:
          type: string
        catalogVisibility:
          type: string
          enum: &id002
          - public_listed
          - public_unlisted
          - private
        visibility:
          type: string
          enum: &id003
          - public
          - instructors_only
        isFrozen:
          type: boolean
        frozenAt:
          type: string
          format: date-time
          nullable: true
        frozenByUserId:
          type: string
          nullable: true
        freezeReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdByUserId:
          type: string
          description: 作成者ユーザーID（instructor 作成 or operator/root 作成）
        status:
          type: string
          enum:
          - draft
          - pending_approval
          - active
          - archived
          description: 講座ステータス
        approvalRequestedAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        approvedByUserId:
          type: string
          nullable: true
      additionalProperties: false
      required:
      - visibility
      - updatedAt
      - createdByUserId
      - createdAt
      - title
      - isFrozen
      - ownerUserId
      - catalogVisibility
      - id
      - status
...
    CourseUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ownerUserId:
          type: string
        catalogVisibility:
          type: string
          enum: *id002
        visibility:
          type: string
          enum: *id003
        isFrozen:
          type: boolean
        frozenAt:
          type: string
          format: date-time
        frozenByUserId:
          type: string
        freezeReason:
          type: string
      additionalProperties: false
...
    CourseChannel:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        type:
          type: string
          enum: &id006
          - assignment
          - one_on_one
          - announcement
          - general
        postingMode:
          type: string
          enum: &id007
          - mixed
          - threads_only
        visibility:
          type: string
          enum: *id003
        isFrozen:
          type: boolean
        frozenAt:
          type: string
          format: date-time
          nullable: true
        frozenByUserId:
          type: string
          nullable: true
        freezeReason:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        isManual:
          type: boolean
        systemKey:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - courseId
      - createdAt
      - id
      - isFrozen
      - isManual
      - postingMode
      - type
      - updatedAt
      - visibility
    CourseChannelCreateRequest:
      type: object
      properties:
        courseId:
          type: string
        type:
          type: string
          enum: *id006
        postingMode:
          type: string
          enum: *id007
        visibility:
          type: string
          enum: *id003
        isFrozen:
          type: boolean
        frozenAt:
          type: string
          format: date-time
        frozenByUserId:
          type: string
        freezeReason:
          type: string
        name:
          type: string
        isManual:
          type: boolean
        systemKey:
          type: string
      additionalProperties: false
      required:
      - courseId
      - type
...
    CourseMessage:
      type: object
      properties:
        id:
          type: string
        channelId:
          type: string
        threadId:
          type: string
          nullable: true
        authorCourseMemberId:
          type: string
          nullable: true
        text:
          type: string
        isEmergency:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - channelId
      - createdAt
      - id
      - isEmergency
      - text
      - updatedAt
    CourseMessageCreateRequest:
      type: object
      properties:
        channelId:
          type: string
        threadId:
          type: string
        authorCourseMemberId:
          type: string
        text:
          type: string
        isEmergency:
          type: boolean
      additionalProperties: false
      required:
      - channelId
      - text
...
    FileObject:
      type: object
      properties:
        id:
          type: string
        storageKey:
          type: string
        bucket:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
          format: int32
        checksum:
          type: string
          nullable: true
        isScanned:
          type: boolean
        isMalicious:
          type: boolean
        createdByUserId:
          type: string
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - bucket
      - createdAt
      - fileName
      - id
      - isMalicious
      - isScanned
      - mimeType
      - size
      - storageKey
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
      - success
      additionalProperties: false
...
    CursorPageInfo:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
      required:
      - hasMore
      additionalProperties: false
    OffsetPageInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
          nullable: true
      required:
      - page
      - pageSize
      additionalProperties: false
    ListMeta:
      type: object
      properties:
        page:
          $ref: '#/components/schemas/OffsetPageInfo'
        cursor:
          $ref: '#/components/schemas/CursorPageInfo'
      additionalProperties: false
    UserPublicView:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
      required:
      - id
      - displayName
      additionalProperties: false
    FileObjectView:
      type: object
      properties:
        file:
          $ref: '#/components/schemas/FileObject'
        url:
          type: string
        mimeType:
          type: string
          nullable: true
        size:
          type: integer
          minimum: 0
          nullable: true
      required:
      - file
      - url
      additionalProperties: false
    MessageReactionView:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
          minimum: 0
        reactedByMe:
          type: boolean
      required:
      - emoji
      - count
      - reactedByMe
      additionalProperties: false
    CourseMessageView:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/CourseMessage'
        sender:
          $ref: '#/components/schemas/UserPublicView'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/FileObjectView'
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/MessageReactionView'
      required:
      - message
      - sender
      - attachments
      - reactions
      additionalProperties: false
    CourseChannelSummaryView:
      type: object
      properties:
        channel:
          $ref: '#/components/schemas/CourseChannel'
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        unreadCount:
          type: integer
          minimum: 0
      required:
      - channel
      - unreadCount
      additionalProperties: false
    CourseStatsView:
      type: object
      properties:
        learnerCount:
          type: integer
          minimum: 0
        assignmentCount:
          type: integer
          minimum: 0
        lessonCount:
          type: integer
          minimum: 0
        activeChannelCount:
          type: integer
          minimum: 0
      required:
      - learnerCount
      - assignmentCount
      - lessonCount
      - activeChannelCount
      additionalProperties: false
    CourseSummaryView:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/Course'
        channelCount:
          type: integer
          minimum: 0
        memberCount:
          type: integer
          minimum: 0
        isFrozen:
          type: boolean
      required:
      - course
      - channelCount
      - memberCount
      - isFrozen
      additionalProperties: false
    CourseDetailView:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/Course'
        channels:
          type: array
          items:
            $ref: '#/components/schemas/CourseChannelSummaryView'
        stats:
          $ref: '#/components/schemas/CourseStatsView'
      required:
      - course
      - channels
      - stats
      additionalProperties: false
...
    CourseListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CourseSummaryView'
        meta:
          $ref: '#/components/schemas/ListMeta'
      required:
      - items
      additionalProperties: false
    CourseChannelListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CourseChannelSummaryView'
        meta:
          $ref: '#/components/schemas/ListMeta'
      required:
      - items
      additionalProperties: false
...
    MessageListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CourseMessageView'
        meta:
          $ref: '#/components/schemas/ListMeta'
      required:
      - items
      additionalProperties: false
...
    GenericWriteRequest:
      type: object
      additionalProperties: true
    GenericDetailView:
      type: object
      additionalProperties: true
...
    CourseChannelDetailView:
      type: object
      allOf:
      - $ref: '#/components/schemas/CourseChannel'
    InstructorCourseCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        catalogVisibility:
          type: string
          enum:
          - public_listed
          - public_unlisted
          - private
        visibility:
          type: string
          enum:
          - public
          - instructors_only
      additionalProperties: false
      required:
      - title
...